Move zen entry from activity bar to agent header
The zen toggle in the layout activity bar is rarely used. Merge it into the
agent header as a special button next to the edit-agents pencil, mirroring
that pattern. The exit-zen button remains in place inside zen mode.
